# packages/proxy/src/openapi.yaml
openapi: 3.1.0
info:
  title: HAAI API
  description: |
    Management REST API and OpenAI-compatible inference endpoints for HAAI.

    **Management API** (`/api/v1/*`) requires admin authentication via session cookie or Bearer token.

    **Inference API** (`/v1/*`) requires an API key in the `Authorization: Bearer` header.
  version: 0.0.1
servers:
  - url: /
    description: Current host
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: haai_session
    bearerAuth:
      type: http
      scheme: bearer
    apiKeyAuth:
      type: http
      scheme: bearer
      description: Client API key (haai-sk-…)
...
paths:
...
  /v1/chat/completions:
    post:
      tags: [OpenAI]
      summary: Chat completion
      security: [{ apiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model, messages]
              properties:
                model: { type: string }
                messages:
                  type: array
                  items:
                    type: object
                stream: { type: boolean }
      responses:
        "200":
          description: Completion response or SSE stream
...